chore: rename recordDecorators to recordModifiers
