+++ src/test/java/edu/kit/datamanager/pit/recordModifiers/CopyAttributeTest.java
@@ -1,4 +1,4 @@
-package edu.kit.datamanager.pit.recordDecoration;
+package edu.kit.datamanager.pit.recordModifiers;
 
 import edu.kit.datamanager.pit.domain.PIDRecord;
 import org.junit.jupiter.api.BeforeEach;
